Prevent memory leaks
The builder.containers list creates circular references between itself and the
ObjectBuilder object that contains it. This means that garbage collection is
taking care of disposing of these self-referencing objects from time to time.
If no GC is enabled (e.g., when using timeit.timeit, or because intentionally
turned off) these objects build up in memory, defeating the purpose of the
items generator and those under it. By explicitly clearing up the builder's
list before the builder gets out of scope we solve this.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>