Fixed path when map_key is first element
gen->path always contains at least 1 element, and when a map_start event is
found a new None element is added to the list. Thus the logic to find whether
this map key is the first element of the prefix returned by this generator was
wrong and generated ".key" for a JSON like '{"key": 1}'

I've added a unit test to make sure this problem doesn't arise again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>