Removing unnecessary conversion to double
When returning Decimals from the C extension we were unnecessarily
turning the string given to us by yajl into a double first, and feeding
that into Decimal, thus completely missing the point of having a Decimal
in the first place.

Fixes #1, even though it was "empty" PR apparently created with the sole
purpose of notifying about the bug.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>