Offering better backend by default
Instead of always importing the python backend and exposing its
functions as the main functions of ijson, let's loop through all
available backends in performance order, and try to use those first. If
they fail to import that's fine as the python backend will still get a
change. If no backend makes it then that's an error.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>