Actually test the C backend
Python always takes the cwd as the first lookup place for
packages/modules. Due to this, when running the unit tests the local
version of the code (that didn't have the C extension compiled) was run,
and thus it has not been included in the tests so far (because the
_yajl2 module cannot be loaded, as it doesn't exist locally).

This commit changes the steps we follow before running the unit tests.
Instead of installing the package via pip we now build it and install it
in development mode via setuptools and then run the tests, ensuring the
C extension is available.

I'm also taking the chance to improve the unit test running itself,
using pytest for a more clear output.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>