Only increase reference if using object.
Otherwise when map_type is None it will still have a reference added
into it, and although it's not obvious if this is a bad thing in
practice, in principle one wants to balance out reference counting.

The other option to conserve balance is to always increment the
reference of the incoming map_type, but then also always decrease it
later, which means we'd need to store None as the value of
builder->map_type. This isn't too bad, but then when checking if we have
a map_type we'd have to change from "if (builder->map_type)" to "if
(builder->map_type != Py_None)", which is slightly less efficient.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>