New function for automatic reader conversion
All ijson backends, except for the python backend, need a bytes reader
in order to work. With the change of default backends in ijson 2.5 it
became clear that many people relied on the ability of the python
backend to accept string readers, and thus ijson's public API broke.

This commit adds a new function to turn a utf8-encoded string reader
into a bytes reader. It is mostly intended for use within ijson, but
users might find some of it useful.

Related issues: #12, #13

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>