Using new automatic reader conversion in backends
All backends, except for the python backend, require a bytes reader, but
users didn't expect this and some were give us a utf8-encoded string
reader instead. With this commit we check and automatically translate
the given reader so users don't have to bother (yet) doing it. Together
with the translation a DeprecationWarning is issued to make it clear
that this automatic translation won't be there for long.

Related issues: #12, #13

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>