Improving error handling in yajl2 backend
In ctypes in general it appears not to be possible to propagate an
exception originating from a callback (e.g., KeyboardInterrupt) back
into the main python code that initiates the call to the C library. This
solution tries to capture these and convert them into a proper return
code. Note however that there's still a small window of opportunity for
a new exception to be raised.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>