Catching only numeric conversion errors
The new set of tests introduced in the previous commit highlighted a
corner case that wasn't properly covered, and that was mentioned in #16:
the yield expression within the try/expect block can also actually raise
an exception if the generator is being destroyed before it is consumed
fully (e.g., if manually iterated over via next() and then discarded).
In these cases an UnexpectedSymbol exception would be raised during the
destruction of the generator object, which would be ignored by the
interpreter.

The solution is simple: include *only* the number parsing logic within
the try/catch, leaving the yield expression outside to get a normal
logic flow.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>