Avoiding doctest until failure cause is corrected
Adding the doctest execution in Travis uncovered an issue in most of the
backends when working with python 2.7: object keys are not returned as
unicode objects. The built-in json module returns unicodes for both
object keys and string values, and what we should do that as well --
otherwise we can't properly support non-ascii key names, which are
currently returned as encoded bytes to the user (not nice!).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>