python2: JSON key names as unicode, not str
JSON key names are allowed to take almost any character, just like
string values. However, most backends were not returning a unicode
object in python2, but a str object instead containing the utf-8 encoded
bytes of the key name. This is the wrong thing to return, and in fact
the built-in json module returns unicode objects in all cases (which can
still be safely compared to str objects in most cases).

The fact that these backends did not return unicodes seems to be a
historical oversight, as the original b2s() calls were in place from the
first commit introducing 2/3 compatibility (see 56234d6) and were never
modified after that. In the case of the C backend this was definitely my
error. The python backend on the other hand was the only "good" one.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>