Declare we use Py_ssize_t, not int
This applies to calls to Py_BuildValue and others that take both a
pointer and a size, and take a format string like "s#", "y#" and so on.
In particular we call Py_BuildValue("(#s)") when formatting Decimal
values out of strings. Setting the macro lets python know whether we are
using a Py_ssize_t or an int to hold the size.

We have always used Py_ssize_t (well, almost... we use the size_t coming
from yajl directly), but on the other hand we never *declared* we were
using it, when in effect we should have done so. In python 3.8 the old
"int" type is being deprecated, and hence the macro is always required.
In 3.9 or 3.10/4.0 support for int sizes will be removed, and the macro
will not be needed anymore.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>