Encapsulating test class creation
Firstly, this avoid polluting the module's namespace, but it also makes
this code reusable if we want to apply a similar approach for other
classes (which we will in the future, once we implement the new
asyncio-enabled generators).

Also, let's move from the *Parse suffix to *Tests, which is a more
meaningful name.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>