Making utils useful again
Until now the utils module hasn't seen any usage in ijson, and thus its
code can be considered dead. However, with the new design we are
planning for turning generators into coroutines, we will start using it
again. Or at least the coroutine decorator will be useful, the rest can
go. Taking the place of the removed functions are new ones that will aid
us on creating the coroutine pipelines we want to offer with ijson, as
well as code turning these new coroutine pipelines into the old-style
generators.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>