Adding new string_reader utility function
This is the inverse function of bytes_reader(), as it ensures the
file-like object reads (unicode) strings, not bytes. This check is
currently done by the python backend, which works on unicode strings
rather than bytes, but since we are planning to remove I/O from the
generators (and turn them into coroutines) the check now needs to be
somewhere else.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>