New parse, kvitems and items coroutines
These have a _coro suffix in their names to distinguish them from the
originals. Although they are almost exact copies from the corresponding
generator, they receive values and dispatch their results to a target
(which should be a coroutine-like object) rather than pull values from
another generator and yield their results.

Just like the generators contained in this module, these coroutines are
meant to be reused by almost all backends (with the exception of the C
backend, which will re-implement them in the name of performance).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>