python backend ported to be coroutine-based
Since this backend implements a lexer and a parser, it was a bit more
difficult to port to be coroutine-based. The main problem was that as
previously implemented, the backend had recursive calls into the parsing
functions, which doesn't work in a coroutine-based world. Hence we have
to translate the parsing functions into a single coroutine, where state
is explicitly kept in the coroutine instead of relying on the call stack
to keep it for us.

Differently from the other backends, this one relies on receiving
unicode strings rather than byte strings, and needs to signal this
accordingly when calling enrich_backend.

Because of all these differences, it doesn't make sense to keep the old
Stream tests anymore. They could potentially be rewritten, but it
doesn't make too much sense, given that the python backend is always
unit tested anyway.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>