New, broken-down C backend implementation
There is simply no way of breaking this into smaller commits that make
sense. Through a long, iterative process we took the code that used to
be in a single source code file, and broke it down into individual
pieces that made sense, while at the same time adding the coroutine
functionality at each level.

The result of this process is revamped C extension implementation. The
main points to understand it are:

 * The most basic types are the *_basecoro coroutines. Unlike the other
backends, all four coroutines are implemented instead of just
implementing basic_parse_basecoro and reusing the common python
implementation. The latter is possible (and worked during development),
but in the name of performance it's better to have everything in C.
 * When dispatching results from one coroutine to another we check the
type of the target coroutine, and if it's one from the same C backend
then we avoid tuple packing/unpacking, plus the dynamic call nature of
calling generic functions in python, even at the C API level. Instead,
we call directly their *_send_impl() C functions, which is quite more
performant.
 * File-reading-based generators share their common file reading and
iteration logic in the form of the reading_generator_t type and
associated functions. Each generator's init function initialises its
internal reading generator object with the corresponding coroutine
pipeline, which is then internally (in the reading_generator_t type)
wired to the file reading iteration process, and a list where final
events are appended and yielded over, very much like in the python
implementation.

A few other cleanups went in this effort, like not requiring to pass the
ijson exception types and the Decimal type from python down to the C
extension (that was just laziness before), apart from the fact that now
all sources are contained in a new sub-directory for better organising
them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>