yajl2_c backend ported to be coroutine-based
The new C extension provides an implementation for each of the four base
coroutines and generators, and is simpler to integrate into the
wrapping python module, which we had to slightly modify to expose all
the underlying functionality. By switching which C sources make up the
extension, plus changing the thin python module that calls into it, we
effectively change the implementation of the yajl2_c backend.

This new broken-down C extension has now header files that are not
automatically included in source distributions; let's add those to the
MANIFEST file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>