New async iterable versions of generators
When using python 3.5+ we further enrich a backend by adding a set of
*_async asynchronous iterables (usable in "async for" loops)
corresponding to each of the four coroutines offered by ijson. This
means we now have built-in support for "async for" iteration in asyncio
event loops, given that the file-like object can be asynchronously read
of course.

The way we iterate over the results in our async_iterator class means we
pop them out from the beginning of the collection rather than the end,
which for lists is a costly operations requiring moving all remaining
elements in memory. We thus use a deque instead, with an additional
send method (a la sendable_list). An alternative would have bee to use
sendable_list and then iterate over the results, keeping track of which
had been returned to eventually empty the list and start again (like we
do in coros2gen), but after some experimentation it turned out it didn't
perform as well.

Like with the coroutines, we are also making these async iterators
available at the top-level ijson module.