New pre-chained, user-facing coroutines
Now that we implemented coroutines for all modules, we can expose them
(in addition of the old generators) at the top-level ijson module from
the default backend. Users should not be offered the low-level
coroutines implementing the different parsing bits of logic though; if
they wanted to use them they would have to chain them themselves, making
it a bit awkward to work with them. Instead we offer pre-chained
co-routines that users can invoke directly, which receive data chunks on
one end, and spit the corresponding event on the other.

These user-facing coroutines are named *_coro, which is why the actual
coroutines implementing the logic bits are called *_basecoro.

Also, since we are giving users the ability to work with coroutines
directly, we should also make it easier for them to prime *their*
target coroutines. For this we are now exposing the utils.coroutine
decorator at the top-level ijson module for easy import in user code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>