Adding notes about old and new design
This will hopefully be a good source of documentation for people wanting
to play with ijson in the future. They explain the old pull pipeline,
and how the new design of a push pipeline actually offers more
flexibility and allows for more functionality to be offered out of the
box.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>