ijson 3.0rc1
This is the first publicly released version of ijson containing the new
underlying design for the parsing logic, plus async support out of the
box. This is going out in the form of a release candidate mainly to
allow users to give some feedback before doing a final release, but also
because there are still known gaps in the code that need to be
addressed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>