Fixing double priming of coroutines
User-facing coroutines had a bug in them as they were being primed
twice; i.e., next() was being called *twice* on them instead of only
once, meaning that the second call produced a None buffer being sent to
the underlying *_basecoro objects.

This was caught while trying to adjust the unit tests that are currently
in place for the old generators, and trying to reuse them for the
coroutines as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>