Complete revamping of unit tests
Because old versions of ijson offered only generators, all unit tests
were written using that method to iterate over JSON content and do
certain checks. We had already written some async-specific basic unit
tests, but they didn't come even close to the amount of coverage the
original set provided.

This commit completely redesigns the way we do unit testing over the
different methods offered by the library. There is a base class that
contains all tests that are common to all methods (coroutines,
generators and async generators), which is combined with a different set
of classes implement the value accumulation logic for each type of
parsing iterator (Coroutine, Generator and Async) to offer the same
suite of tests for all iterators. On top of that, we still generate
individual test classes for each backend. This de-coupling of parsing
accumulation and the corresponding checks makes most of the tests valid
for all methods, and ensures we actually test our new coroutines and
async generators as thoroughly as we did our old generators.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>