Simplifying test execution
Because tests for the async iterators are in a different module, so far
we have required users to run both tests.py and tests_async.py to run
the whole test suite. Although a way to automate this is having an
actual test/ package and have auto-discovery of tests, etc., a different
way is to import tests_async from tests and generate the Async* classes
inside the tests module directly. This way users only need to execute
tests.py, which simplifies things quite a bit in general, specifically
to achieve better coverage with tox, and to simplify our Travis CI
script.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>