Removing unused callbacks
When the yajl_number callback is defined, yajl_integer and yajl_double
are ignored. All callbacks have defined these two callback functions
anyway, and therefore our code coverage is suffering because of this.

This commit removes unused callbacks in the yajl2_cffi and yajl2_c
backends, where they are defined as stand-alone functions that are not
invoked. In the yajl and yajl2 backends the situation is different (they
are defined as lambdas as part of a list of tuples) and it's more
difficult to remove them, so instead of removing them altogether, we
remove the logic inside of them to make it more explicit how they do
*not* add value to the parsing process.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>