Changing checks by assertions
These checks never really yielded false because they were meant to
double-check we were doing things correctly internally in the C backend
-- in other words, they are asserts and not validations, and therefore
should be expressed as such. This has obviously the nice effect it
increments the coverage of the C backend code.

Finally, an error would have already been raised when getting the 'read'
attribute from the file-like object in the C extension if the given
object didn't have any, so there's no point on having a final "else"
statement; even in the case of users not passing down a file-like object
an error would have been raised.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>