Improving coroutine protocol for users
If users decide to use the user-facing coroutines directly previously
needed to feed manually a final empty bytes string (or unicode string in
the case of the python backend) to trigger the finalisation of the
parsing process. This was unintuitive and didn't play well with the
protocol generators/coroutines offer out of the box in python. A better
solution, implemented in this commit, is to offer users the opportunity
to call the close() method once they are finished, which will internally
trigger the final iteration on the coroutine and provoke them to exit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>