Adding C coverage measurements
Coveralls.io doesn't allow multiple uploads from the same job, so
instead we use one of the jobs to upload the C measurements instead of
the python measurements. The one job we chose for this doesn't add
anything by itself to the python measurements, so we are only winning
here.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>