Supporting str literals in self.all/self.first
When users give a string stream (as opposed to a bytes stream) to one of
the methods accepting a file-like object, almost all backends will
perform automatic encoding of the stream on-the-fly, while emitting a
warning about it. We had a test for this situation with the original
methods, but because of how it was written it was impossible to reuse it
for the *_async methods, which so far haven't been tested against this
(but should work).

This change adds the necessary infrastructure to make the existing test
reusable across the async and non-async tests.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>