Making chain() function reusable
This will be used by the new implementation of the async iterables in
the C backend, who will also need to chain together the different
basecoro objects to form a coroutine pipeline.

I took the chance to also rename the structure holding the information
to create these chains to pipeline_node, which is a bit more
informative and makes the code a tiny bit more readable.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>