Introducing the new async_reading_generator type
The implementation of any given asynchronous iterable (like
basic_parse_async, parse_async, etc.) requires us to define a type with
two __aiter__ and __anext__ methods. The first method simply returns
self, but the second has to return an *awaitable* each time it is
called, In C awaitables can be implemented as *iterables* (a type with
__iter__ and __next__), with each invocation to __next__ corresponding
to a resume on the logic of the awaitable, and a final StopIteration
exception with a value signifying that a value is being returned by the
awaitable. In other words, this iterable is the one actually doing the
logic usually found in __anext__ in python classes, or (in python 3.7+)
in asynchronous generators.

All asynchronous iterables we need to offer have the same logic: they
read data off a file, and pass down the events through a pipeline of
coroutines, which eventually fills a list of events, where we finally
yield them from. This common code can be abstracted away from the
individual asynchronous iterables, and offered as an iterable so that
individual asynchronous iterables can return them in *their* __anext__
methods. The async_reading_generator is this abstraction.

Just like its python counterpart (the utils35.async_iterable class),
this iterator performs two different "await" operations: one for
ensuring a read() method returning bytes is used throughout the process,
and then N for reading data off the read() awaitable. As data is read
and parsed, a list is filled with events, from which values are returned
by continuously raising StopIteration. Finally, when no more reading can
be done, and no more events are there to be returned, an
StopAsyncIteration is raised.

This type is very much like the existing reading_generator, except that
because it's a user-facing type it needs to be properly registered
instead of operating in the background. After construction each
asynchronous iterable also needs to indicate what is the coroutine
pipeline that needs to be constructed so the correct flow of events
occurs through the system.

Note that this commit doesn't include any asynchronous iterable
implementation yet, but only the underlying awaitable/iterable that will
underline them all. The asynchronous iterables will be added in future
commits.

Heavy inspiration was drawn from [1] and [2] (both from the same person)
in order to fully understand the await/async protocol at the C level.

[1] https://stackoverflow.com/a/51115745/7108753
[2] https://stackoverflow.com/a/54675255/7108753

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>