Removing very-improvable "if"s
We are not really removing them, but replacing them with the macros we
use across the whole C extension. The only difference is that the buffer
objects won't see their reference counts decreased, but: a) the chances
of this error occurring are actually very low (if not nil?), and b) the
size of buffer shouldn't be too big, so users shouldn't be heavily
impacted.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>