New ability to build ijson including yajl sources
Having a "fat" build of the C extension that includes the yajl sources
is an easy way to be able to build a binary wheel without external
dependencies, and that therefore doesn't need to be "fixed" later on by
embedding the external library using audit-wheel, delocate-wheel or
similar tools. This, in turn, opens the door to build binary wheels for
Windows, where such wheel patching looks to be problematic (or at least
that's the impression I got from reading on Internet).

The source code layout that is needed by for these changes to work is
quite specific, and hence I don't really expect users to be using it
directly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>