Adding AppVeyor support
By adding AppVeyor we now can build binary ijson wheels for Windows.
Older versions of VisualStudio didn't support C99's designated
initialisation, so unless we change the code to fully describe each
PyTypeObject in full we won't be able to support versions < 3.5. This is
however a good compromise for this first version, so unless users ask
for it I will not be putting effort into adding such support.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>