Reintroduce common.* functions
This commit re-introduces the common.* functions that were removed
during the redesign of ijson 3. In ijson 2.X these functions were the
building blocks over which the backend-specific functions were built
upon. However, with the redesign of the library in version 3 they were
not needed anymore and they were removed.

It turns out users around the globe used these methods, as they iterated
over events in ways we didn't anticipate. Thus, there is value in adding
these functions back, even if it's not in a fully-performant way. With
time we can think of moving these functions into each backend so they
use the backend's underlying function, which is important for
performance.

Tests have been added to double-check that the behavior is the expected
one. In the case of items and kvitems we event skip through a few events
before reading the rest, which cannot be safely done with parse, as its
state machine depends on reading all events in order.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>