Port python backend to work with bytes
The python backend was the only one that internally still required
strings (and therefore a string reader) rather than bytes, which
complicated things a bit around testing and overall maintenance.

With this change we are changing how the python backend works, now
dealing with bytes rather than strings, and therefore requiring a bytes
reader just like the rest of the backends. To be as least invasive as
possible we implemented this as an additional coroutine that comes
before the Lexer and incrementally decodes bytes into UTF-8 strings;
these strings are are then passed down to the rest of the machinery,
which doesn't require any further changes.

Benchmarks with python 3.8 show no change in performance before and
after this change (only within noise).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>