React uniformly to incorrect UTF-8 bytes
The way the library has so far reacted to invalid UTF-8 byte sequences
has not been ideal, as different types of errors are raised by different
backends, with more or less information about the underlying reason of
the failure.

With this commit invalid UTF-8 byte sequences are now always reported as
IncompleteJSONError exceptions. In the case of the yajl2-based backends
we also take care of constructing the exception with a bytes() object
instead of a str() when failing when building the error message (because
of the decoding that takes place). Previously these message conversions
meant that UnicodeDecodeError exceptions were raised, making it less
clear to users the context in which the invalid byte sequences were
found, which is more obvious from the yajl error message.

Thanks to this new uniformity we can now actually test that invalid
UTF-8 byte sequences are rejected with a JSONError.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>