Remove internal "use_string_reader" flag
This flag was used by the python backend to signal that it worked with
string readers rather than byte readers, which in turn made a few things
work differently. Now that the python backend works with byte streams
(in hindsight it was way easier to implement the latter...) there is no
need to keep the code supporting this feature around.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>