Remove memory leak in yajl2_c's parse_basecoro
This problem was initially reported in #28. Since memory leaks are still
checked manually (i.e., there is no automatic unit tests that check
memory consumption remains constant throughout an ijson execution) this
leak was somehow missed until now.

The problem was *not* present in the 2.x series, and this Py_DECREF()
call got lost during the refactoring for version 3.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>