Fix items and kvitems to work with empty members
The implementation of items and kvitems, both their pure python and in C
variants (the latter being based on the former), used the path coming
from the parse() events as part of the conditions used to enter the
object building and accumulation logic, and to finish the building of a
particular object (and communicate it furthern down the pipeline). This
is possible as long as one assumes the path always changes when entering
into a new object or array in the JSON document.

Alas, this change in path does not *always* happens. The exception to
this is when the top-level JSON value is an object with an empty-named
member (e.g., {"": 1}). In these cases the events generated from parse()
will contain an empty path for the top-level object itself and *also*
for the events generated in the context of the empty member (in the
example above, when communicating the value "1" parse() will generate a
(path, event, value) tuple == ("", "number", 1). This exception breaks
the assumptions made so far by the items and kvitems implementations,
having different unexpected results: the python kvitems implementation
raised an AttributeError due to invalid usage of the ObjectBuilder
object, the C kvitems implementation crashed with a segmentation fault,
and both the python and C items implementations produced wrong results
in particular cases.

This commit changes the implementation of the kvitems() and items()
functions, both in python and in C, using the same logic. In all cases
in order to check if the building of the object is complete we don't
compare the incoming path to the prefix anymore (because we know now it
is not a correct comparison in all cases); instead we manually keep
track of the additional depth introduced by incoming events, and finish
building our objects when that depth goes back to the expected base
value. Although slightly more complex, these new implementations
actually carry out less (or at least shorter) string comparisons, and
hence are slightly faster.

Note that a limitation that is made clear by the "emtpy member at the
top-level" exception is that such elements cannot currently be selected
by a prefix, because empty prefixes refer to the top-level object and
not any of its members. Therefore selecting "" from {"": 1, "a": 2}
yields the whole object instead of just "1", but selecting "a" yields
"2". This is a problem that will need further thinking, but probably
a limitation we can live with anyway.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>