Add support for new use_float option
The new use_float option (defaults to False for backwards compatibility)
instructs the basicparse routines of each parser to return float()
values instead of Decimal.Decimal() values for non-integer numbers found
during the parsing process.

The way we implement this change in the yajl-based backends is by not
providing a "number" callback, in which case the "integer" and "double"
callbacks are called. By then the string has already been parsed by yajl
into C "long long" and "double" values respectively, so the only thing
that is left to us is to wrap those values into their Python objects.
This has a few consequences: first, integer parsing works faster in this
regime because yajl does a better work than us at it; secondly, creating
python float() values out of C doubles is way faster than creating
Decimal() values out of C strings, so again performance increases.
Finally, since yajl does the string->double conversion for us, we have
no control on how precision is lost, and when. In particular, underflow
is fine (very small numbers are returned as 0), but big numbers result
in an overflow error state from which one cannot recover. The python
backend will happily create a float "inf" value from these values
though, so we have to manually check for this and raise an error as well
to keep behavior consistent across all backends.

This loss in precision is probably something most users can deal with
though, and the overflows are rare, so in the future we will probably
change the default value of use_float to True given its better
performance.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>