Rename number function
This function takes a string an returns either an integer or a Decimal
object. We are planning to support an option to return float() values
instead of Decimal objects for non-integer values we will be adding a
similar function, and in this new context the original name will be
confusing.

This function has never been meant to be publicly exposed, but some
users might be using it anyway, so I kept it with a DeprecationWarning
for temporary backwards-compatibility. A new test checks that this is
true.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>