De-duplicate ctypes yajl handling
The yajl and yajl2 backends, both based on ctypes, need very similar
code to register callbacks, retrieve the library, and setup the
environment for correctly calling into the DLL. This code was basically
duplicated across both modules.

This commit removes the duplicated code from each of the modules and
puts it under a new common module that both backends can use to set
themselves up.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>