Add non-concatenating path mode for parse_basecoro
Historically the parse routine has been always creating prefix-like
paths by concatenating the map keys and the "items" literals using "."
as the separator. These values are then given either to the user (when
users call the parse routines directly) or to the kvitems/items
routines. In these latter cases these computed paths are what
kvitems/items compare against to determine when a user-given prefix has
been reached, and thus when a new object needs to be start being built,
and eventually returned to the user.

It turns out there is an alternative way for kvitems and items to
interact with parse. Instead of having parse create concatenations of
the path elements into single strings, which then kvitems/items compare
against the user prefix, we can do something more efficient: first we
simply collect the path elements into a list in parse and communicate
that list of path elements to kvitems/items; they in turn take the
original prefix given by the user and turn it into a list of path
elements. The comparisons done by kvitems/items are then performed on
the equality of the two lists, which internally will involve a size
comparison first, and then an element-wise comparison. This different
mode of operation relieves parse from having to perform all that string
manipulation, avoiding object creation and string handling, therefore
improving overall performance.

This commit only adds the new mode of operation in parse, which still
defaults to the old behavior.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>