Port kvitems/items to use new parse mode
Now that yajl2_c's parse method supports a new mode of operation where
paths are not concatenated anymore we can port kvitems and items to use
this method. This requires only flipping a flag to False, which
otherwise defaults to True for the old behavior.

Because this additional flag that needs to be passed down to parse we
cannot use common._make_kvitems/items_coro functions anymore (i.e.,
because the other parse functions don't have this flag). Even if we
extend to accept an optional parse_args parameter we still have the
issue that the backend they'd need to receive is _yajl2_c (not yajl2_c),
where the different basecoro __init__ functions have slightly different
signatures from the standard ones (to keep the C code simpler). The
simplest solution then is to provide our own pipeline definition for
kvitems and items, keeping it as non-repetitive as possible.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>